Move Figure class to `figureMedia` directory
