+++ packages/11ty/_plugins/figures/figureMedia/factory.js
