+++ packages/11ty/_plugins/figures/figureMedia/index.js
