+++ packages/11ty/_plugins/figures/test/figures.spec.js
@@ -3,7 +3,7 @@ import { dirname, resolve } from 'node:path'
 import { fileURLToPath } from 'node:url'
 import { readFileSync } from 'node:fs'
 import Ajv from 'ajv'
-import Figure from '../figure/index.js'
+import Figure from '../figureMedia/index.js'
 import Manifest from '../iiif/manifest/index.js'
 import assert from 'assert/strict'
 import figureFixtures from './__fixtures__/figures/index.js'
